Test if statment

+++ ci_cd/build_release.py
@@ -23,7 +23,10 @@ for addedfile in newfileslist:
   if directory == 'databricks':
     print(' Added file for databricks is: ' + addedfile)
 
-    shutil.copytree(src, dest)
+    # shutil.copytree(src, dest)
+    if os.path.exists(dest):
+        shutil.rmtree(dest)
+        shutil.copytree(src, dest)
 
     # Pass variables from script to azure devops pipeline
     print('##vso[task.setvariable variable=directory;]%s' % (directory))
